+++ src/Lykke.Service.PayInternal/Models/PaymentRequests/PaymentRequestDetailsModel.cs
@@ -17,7 +17,7 @@ namespace Lykke.Service.PayInternal.Models.PaymentRequests
         
         public string MerchantId { get; set; }
         
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         
         public string SettlementAssetId { get; set; }
         
@@ -34,7 +34,7 @@ namespace Lykke.Service.PayInternal.Models.PaymentRequests
         [JsonConverter(typeof(StringEnumConverter))]
         public PaymentRequestStatus Status { get; set; }
         
-        public double PaidAmount { get; set; }
+        public decimal PaidAmount { get; set; }
         
         public DateTime? PaidDate { get; set; }
